add code to prevent non-superusers access

+++ products/views.py
@@ -3,6 +3,7 @@ from django.shortcuts import render, redirect, reverse, get_object_or_404
 from django.contrib import messages
 from django.db.models import Q
 from django.db.models.functions import Lower
+from django.contrib.auth.decorators import login_required
 from .models import Products, Category
 from .forms import AddProductForm
 
@@ -73,8 +74,14 @@ def product_details(request, product_id):
     return render(request, 'products/product_details.html', context)
 
 
+@login_required
 def add_product(request):
     """Add a product to the store - Super User Only"""
+    if not request.user.is_superuser:
+        # Redirect back to homepage if not a superuser
+        messages.error(request, 'Sorry, only store owners can do that.')
+        return redirect(reverse('home'))
+
     if request.method == 'POST':
         form = AddProductForm(request.POST, request.FILES)
         if form.is_valid():
@@ -94,8 +101,14 @@ def add_product(request):
     return render(request, template, context)
 
 
+@login_required
 def edit_product(request, product_id):
     """Edit a product on the store - Super User Only"""
+    if not request.user.is_superuser:
+        # Redirect back to homepage if not a superuser
+        messages.error(request, 'Sorry, only store owners can do that.')
+        return redirect(reverse('home'))
+
     product = get_object_or_404(Products, pk=product_id)
     if request.method == 'POST':
         form = AddProductForm(request.POST, request.FILES, instance=product)
@@ -120,8 +133,15 @@ def edit_product(request, product_id):
 
     return render(request, template, context)
 
+
+@login_required
 def delete_product(request, product_id):
     """ Delete a product from the store """
+    if not request.user.is_superuser:
+        # Redirect back to homepage if not a superuser
+        messages.error(request, 'Sorry, only store owners can do that.')
+        return redirect(reverse('home'))
+
     product = get_object_or_404(Products, pk=product_id)
     product.delete()
     messages.success(request, f'{product.name} deleted!')
